+++ src/config.php
@@ -53,6 +53,23 @@ return [
     // to be larger than the original source image
     'allowUpScaledImageVariants' => false,
 
+    // Controls whether images scaled down >= 50% should be automatically sharpened
+    'autoSharpenScaledImages' => true,
+
+    // Default aspect ratios
+    'defaultAspectRatios' => [
+        ['x' => 16, 'y' => 9],
+        ['x' => 8, 'y' => 5],
+        ['x' => 4, 'y' => 3],
+        ['x' => 5, 'y' => 4],
+        ['x' => 1, 'y' => 1],
+        ['x' => 2, 'y' => 2],
+        ['x' => 9, 'y' => 16],
+        ['x' => 5, 'y' => 8],
+        ['x' => 3, 'y' => 4],
+        ['x' => 4, 'y' => 5],
+    ],
+
     // Default image variants
     'defaultVariants'            => [
         [